requirements.txt
# Connect Smart - Dependencies

# Core AI Framework
langgraph>=1.0.0
langchain>=1.2.0
langchain-anthropic>=1.3.0
langchain-community>=0.4.0

# Browser Automation
browser-use>=0.11.0
playwright>=1.58.0

# Voice AI
httpx>=0.28.0

# Web Framework
fastapi>=0.128.0
uvicorn[standard]>=0.40.0
python-multipart>=0.0.12

# Telegram
python-telegram-bot>=22.0

# WhatsApp
aiohttp>=3.10.0

# Database
supabase>=2.27.0

# Utilities
python-dotenv>=1.2.0
pydantic>=2.12.0
pydantic-settings>=2.12.0

# Async support
tenacity>=9.0.0

# Timezone
pytz>=2024.1

# Speech-to-Text
openai>=1.0.0

# Scheduler voor herinneringen
apscheduler>=3.10.0

# Google APIs (Calendar + Gmail)
google-api-python-client>=2.0.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
